+++ main.py
@@ -1,3 +1,5 @@
+print("Launching SweetSuite...", flush=True)
+
 from datetime import datetime
 import logging
 import os
@@ -63,9 +65,6 @@ def apply_light_palette(app: QApplication) -> None:
 def main():
     # Install Qt message handler to suppress warnings.
     qInstallMessageHandler(suppress_qt_warnings)
-    # Show loading screen in case of .exe file.
-    if getattr(sys, "frozen", False):
-        import pyi_splash
     # Setup global logging.
     setup_logging()
     logging.info("SweetSuite application started\n")
@@ -76,9 +75,6 @@ def main():
     # Create and show the main window.
     main_window = MainWindow()
     main_window.show()
-    # Close loading screen in case of .exe file.
-    if getattr(sys, "frozen", False):
-        pyi_splash.close()
     # Start application's event loop, exit when loop ends.
     sys.exit(app.exec())