+++ sweet_suite/mass_spectrometry/analyte.py
@@ -13,11 +13,10 @@ class Analyte:
             `mz_exact`, `relative_area_theoretical`, `area`, `maximum_intensity`
             and `mass_error_ppm`.
         background_and_noise (tuple): Tuple containing the following 
-            information for the monoisotopic peak of the analyte: 
+            information for the lowest-mass peak of the analyte: 
             `(average background intensity, background area, noise)`.
         use_peak_height (bool): If True, use maximum intensity instead of
             trapezoidal area for quantitation.
-        mz_monoisotopic (float): Monoisotopic m/z value.
         isotopic_fraction (float): Theoretical fraction of the isotopic 
             pattern that was integrated.
         mass_error_ppm (float): Mass error in parts per million (ppm), based on 
@@ -53,7 +52,7 @@ class Analyte:
                 `relative_area_theoretical`, `area`, `maximum_intensity`,
                 `mass_error_ppm`.
             background_and_noise: A tuple containing the following information
-                for the monoisotopic peak of the analyte:
+                for the lowest-mass peak of the analyte:
                 `(average background intensity, background area, noise)`.
             use_peak_height: If True, use the maximum intensity of each
                 isotopic peak instead of the trapezoidal area for quantitation.
@@ -65,7 +64,6 @@ class Analyte:
         self.peaks = peaks
         self.background_and_noise = background_and_noise
         self.use_peak_height = use_peak_height
-        self.mz_monoisotopic = self.get_mz_monoisotopic()
         self.isotopic_fraction = self.get_isotopic_fraction()
         self.mass_error_ppm = self.get_mass_error_ppm()
         self.total_area = self.get_total_area()
@@ -74,10 +72,6 @@ class Analyte:
         self.total_area_background_subtracted = self.get_total_area_background_subtracted()
         self.signal_to_noise = self.get_signal_to_noise()
         self.isotopic_pattern_quality = self.get_isotopic_pattern_quality()
-
-    def get_mz_monoisotopic(self) -> float:
-        """Return the exact m/z value of the monoisotopic peak."""
-        return self.peaks.iloc[0]["mz_exact"]
 
     def get_isotopic_fraction(self) -> float:
         """Return the fraction of the isotopic pattern that was integrated."""
@@ -128,7 +122,7 @@ class Analyte:
     def get_total_noise(self) -> float:
         """Return the total noise for the analyte.
         
-        The noise of analyte is calculated as the noise of the monoisotopic
+        The noise of analyte is calculated as the noise of the lowest-mass
         peak multiplied by the total number of isotopic peaks.
         """
         total_noise = self.background_and_noise[2] * len(self.peaks)
@@ -141,7 +135,7 @@ class Analyte:
     def get_total_area_background_subtracted(self) -> float:
         """Return the background subtracted total area (or peak heights).
         
-        In area mode: the monoisotopic background area is subtracted from the
+        In area mode: the lowest-mass background area is subtracted from the
         area of each isotopic peak.
         In peak-height mode: the average background intensity is subtracted
         from the maximum intensity of each isotopic peak.